test_search: Fix Errno::ENOENT for Linuxbrew
Fix error:
IntegrationCommandTests#test_search: Errno::ENOENT:
No such file or directory - /tmp/homebrew_tests/cache/desc_cache.json

+++ Library/Homebrew/test/test_integration_cmds.rb
@@ -666,6 +666,6 @@ class IntegrationCommandTests < Homebrew::TestCase
 
     assert_predicate desc_cache, :exist?, "Cached file should exist"
   ensure
-    desc_cache.unlink
+    desc_cache.unlink if desc_cache.exist?
   end
 end
